BUG-FIX: El operador == usado entre objetos Integer se comporta diferente en las distintas maquinas virtuales, en el caso de la jvm de sun funciona transformando Integer a int pero en el caso de la openjdk 1.6 compara como objetos generando un comportamiento anómalo en el caso de uso de la selección del tamaño de monitor al abrir el llamador. Se corrige usando el metodo Integer.intValue() antes de realizar la comparación. BUG encontrado y resuelto por Adriana Redín del MTSS

+++ SAE-WEB/src/uy/gub/imm/sae/web/common/TipoMonitor.java
@@ -63,7 +63,7 @@ public enum TipoMonitor {
 	public static TipoMonitor fromPulgadas(Integer pulgadas) {
 		if (pulgadas != null) {
 			for (TipoMonitor t : TipoMonitor.values()) {
-				if (pulgadas == t.pulgadas) {
+				if (pulgadas.intValue() == t.pulgadas.intValue()) {
 					return t;
 				}
 			}
